+++ 0x08-recursion/main.h
@@ -17,9 +17,9 @@ int factorial(int n);
 
 int _pow_recursion(int x, int y);
 
-int _sqrt_recursion(int n);
+int _sqrt_recursion(long int n);
 
-int _sqrt_recursion_h(int n, int s, int e);
+int _sqrt_recursion_h(long int n, int s, int e);
 
 
 #endif
